Split custom-element.js into registry, static layout and element class modules
