+++ src/service/extensions-impl.js
@@ -27,7 +27,7 @@ import {
   copyElementToChildWindow,
   stubElementIfNotKnown,
   upgradeOrRegisterElement,
-} from '../custom-element';
+} from './custom-element-registry';
 import {cssText} from '../../build/css';
 import {dev, rethrowAsync} from '../log';
 import {getMode} from '../mode';
