+++ test/functional/test-custom-element.js
@@ -14,152 +14,15 @@
  * limitations under the License.
  */
 
-import * as lolex from 'lolex';
-import {AmpDocSingle} from '../../src/service/ampdoc-impl';
 import {AmpEvents} from '../../src/amp-events';
 import {BaseElement} from '../../src/base-element';
 import {ElementStub} from '../../src/element-stub';
 import {LOADING_ELEMENTS_, Layout} from '../../src/layout';
-import {createElementWithAttributes} from '../../src/dom';
-import {installDocumentStateService} from '../../src/service/document-state';
-import {poll} from '../../testing/iframe';
 import {ResourceState} from '../../src/service/resource';
 import {Services} from '../../src/services';
-import {
-  copyElementToChildWindow,
-  createAmpElementProto,
-  getElementClassForTesting,
-  registerElement,
-  resetScheduledElementForTesting,
-  stubElementIfNotKnown,
-  stubElementsForDoc,
-  upgradeOrRegisterElement,
-} from '../../src/custom-element';
-
-
-describes.realWin('CustomElement register', {amp: true}, env => {
-
-  class ConcreteElement extends BaseElement {}
-
-  let win, doc, ampdoc, extensions;
-
-  beforeEach(() => {
-    win = env.win;
-    doc = win.document;
-    ampdoc = env.ampdoc;
-    extensions = env.extensions;
-    ampdoc.declareExtension_('amp-element1');
-  });
-
-  function insertElement(name) {
-    const testElement = createElementWithAttributes(doc, name, {
-      width: '300',
-      height: '250',
-      type: '_ping_',
-      'data-aax_size': '300*250',
-      'data-aax_pubname': 'abc123',
-      'data-aax_src': '302',
-    });
-    doc.body.appendChild(testElement);
-  }
-
-  it('should go through stub/upgrade cycle', () => {
-    registerElement(win, 'amp-element1', ElementStub);
-    expect(getElementClassForTesting(win, 'amp-element1'))
-        .to.equal(ElementStub);
-
-    // Pre-download elements are created as ElementStub.
-    const element1 = doc.createElement('amp-element1');
-    doc.body.appendChild(element1);
-    expect(element1.implementation_).to.be.instanceOf(ElementStub);
-
-    // Post-download, elements are upgraded.
-    upgradeOrRegisterElement(win, 'amp-element1', ConcreteElement);
-    expect(getElementClassForTesting(win, 'amp-element1'))
-        .to.equal(ConcreteElement);
-    expect(element1.implementation_).to.be.instanceOf(ConcreteElement);
-
-    // Elements created post-download and immediately upgraded.
-    const element2 = doc.createElement('amp-element1');
-    doc.body.appendChild(element1);
-    expect(element2.implementation_).to.be.instanceOf(ConcreteElement);
-  });
-
-  it('should mark stubbed element as declared', () => {
-    expect(ampdoc.declaresExtension('amp-element2')).to.be.false;
-
-    const head = document.createElement('fake-head');
-    const script = document.createElement('script');
-    script.setAttribute('custom-element', 'amp-element2');
-    head.appendChild(script);
-    sandbox.stub(ampdoc, 'getHeadNode', () => head);
-
-    stubElementsForDoc(ampdoc);
-    expect(ampdoc.declaresExtension('amp-element2')).to.be.true;
-    expect(win.ampExtendedElements['amp-element2']).to.equal(ElementStub);
-  });
-
-  it('should install pre-stubbed element extension', () => {
-    const stub = sandbox.stub(extensions, 'installExtensionForDoc');
-
-    stubElementIfNotKnown(win, 'amp-element2');
-    expect(win.ampExtendedElements['amp-element2']).to.equal(ElementStub);
-    expect(ampdoc.declaresExtension('amp-element2')).to.be.false;
-    expect(stub).to.not.be.called;
-
-    const element = doc.createElement('amp-element2');
-    doc.body.appendChild(element);
-    expect(stub).to.be.calledOnce;
-    expect(stub).to.be.calledWithExactly(ampdoc, 'amp-element2');
-  });
-
-  it('should not install declared pre-stubbed element extension', () => {
-    ampdoc.declareExtension_('amp-element2');
-    const stub = sandbox.stub(extensions, 'installExtensionForDoc');
-
-    stubElementIfNotKnown(win, 'amp-element2');
-    expect(win.ampExtendedElements['amp-element2']).to.equal(ElementStub);
-    expect(ampdoc.declaresExtension('amp-element2')).to.be.true;
-    expect(stub).to.not.be.called;
-
-    const element = doc.createElement('amp-element2');
-    doc.body.appendChild(element);
-    expect(stub).to.not.be.called;
-  });
-
-  it('should not install declared pre-installed element', () => {
-    const stub = sandbox.stub(extensions, 'installExtensionForDoc');
-
-    registerElement(win, 'amp-element1', ConcreteElement);
-    expect(win.ampExtendedElements['amp-element1']).to.equal(ConcreteElement);
-    expect(ampdoc.declaresExtension('amp-element1')).to.be.true;
-    expect(stub).to.not.be.called;
-
-    const element = doc.createElement('amp-element1');
-    doc.body.appendChild(element);
-    expect(stub).to.not.be.called;
-  });
-
-  it('insert script for amp-ad when script is not included', () => {
-    insertElement('amp-ad');
-    expect(doc.head.querySelectorAll('[custom-element="amp-ad"]'))
-        .to.have.length(1);
-  });
-
-  it('insert script for amp-embed when script is not included', () => {
-    insertElement('amp-embed');
-    expect(doc.head.querySelectorAll('[custom-element="amp-embed"]'))
-        .to.have.length(0);
-    expect(doc.head.querySelectorAll('[custom-element="amp-ad"]'))
-        .to.have.length(1);
-  });
-
-  it('insert script for amp-video when script is not included', () => {
-    insertElement('amp-video');
-    expect(doc.head.querySelectorAll('[custom-element="amp-video"]'))
-        .to.have.length(1);
-  });
-});
+import {createAmpElementProtoForTesting} from '../../src/custom-element';
+import {poll} from '../../testing/iframe';
+import * as lolex from 'lolex';
 
 
 describes.realWin('CustomElement', {amp: true}, env => {
@@ -249,11 +112,13 @@ describes.realWin('CustomElement', {amp: true}, env => {
     doc.body.appendChild(container);
 
     ElementClass = doc.registerElement('amp-test', {
-      prototype: createAmpElementProto(win, 'amp-test', TestElement),
+      prototype: createAmpElementProtoForTesting(win, 'amp-test', TestElement),
     });
     StubElementClass = doc.registerElement('amp-stub', {
-      prototype: createAmpElementProto(win, 'amp-stub', ElementStub),
+      prototype: createAmpElementProtoForTesting(win, 'amp-stub', ElementStub),
     });
+    win.ampExtendedElements['amp-test'] = TestElement;
+    win.ampExtendedElements['amp-stub'] = ElementStub;
     ampdoc.declareExtension_('amp-stub');
 
     testElementCreatedCallback = sandbox.spy();
@@ -1029,24 +894,24 @@ describes.realWin('CustomElement', {amp: true}, env => {
 
   it('should apply heights condition', () => {
     const element1 = new ElementClass();
-    element1.sizerElement_ = doc.createElement('div');
+    element1.sizerElement = doc.createElement('div');
     element1.setAttribute('layout', 'responsive');
     element1.setAttribute('width', '200px');
     element1.setAttribute('height', '200px');
     element1.setAttribute('heights', '(min-width: 1px) 99%, 1%');
     container.appendChild(element1);
     element1.applySizesAndMediaQuery();
-    expect(element1.sizerElement_.style.paddingTop).to.equal('99%');
+    expect(element1.sizerElement.style.paddingTop).to.equal('99%');
 
     const element2 = new ElementClass();
-    element2.sizerElement_ = doc.createElement('div');
+    element2.sizerElement = doc.createElement('div');
     element2.setAttribute('layout', 'responsive');
     element2.setAttribute('width', '200px');
     element2.setAttribute('height', '200px');
     element2.setAttribute('heights', '(min-width: 1111111px) 99%, 1%');
     container.appendChild(element2);
     element2.applySizesAndMediaQuery();
-    expect(element2.sizerElement_.style.paddingTop).to.equal('1%');
+    expect(element2.sizerElement.style.paddingTop).to.equal('1%');
   });
 
   it('should rediscover sizer to apply heights in SSR', () => {
@@ -1059,10 +924,10 @@ describes.realWin('CustomElement', {amp: true}, env => {
     container.appendChild(element1);
 
     const sizer = doc.createElement('i-amphtml-sizer');
-    expect(element1.sizerElement_).to.be.undefined;
+    expect(element1.sizerElement).to.be.undefined;
     element1.appendChild(sizer);
     element1.applySizesAndMediaQuery();
-    expect(element1.sizerElement_).to.equal(sizer);
+    expect(element1.sizerElement).to.equal(sizer);
     expect(sizer.style.paddingTop).to.equal('99%');
   });
 
@@ -1077,9 +942,9 @@ describes.realWin('CustomElement', {amp: true}, env => {
 
     const sizer = doc.createElement('i-amphtml-sizer');
     element1.appendChild(sizer);
-    element1.sizerElement_ = null;
+    element1.sizerElement = null;
     element1.applySizesAndMediaQuery();
-    expect(element1.sizerElement_).to.be.null;
+    expect(element1.sizerElement).to.be.null;
     expect(sizer.style.paddingTop).to.equal('');
   });
 
@@ -1150,10 +1015,10 @@ describes.realWin('CustomElement', {amp: true}, env => {
   it('should change size with sizer', () => {
     const element = new ElementClass();
     const sizer = doc.createElement('div');
-    element.sizerElement_ = sizer;
+    element.sizerElement = sizer;
     element.changeSize(111, 222, {top: 1, right: 2, bottom: 3, left: 4});
     expect(parseInt(sizer.style.paddingTop, 10)).to.equal(0);
-    expect(element.sizerElement_).to.be.null;
+    expect(element.sizerElement).to.be.null;
     expect(element.style.height).to.equal('111px');
     expect(element.style.width).to.equal('222px');
     expect(element.style.marginTop).to.equal('1px');
@@ -1420,7 +1285,7 @@ describes.realWin('CustomElement Service Elements', {amp: true}, env => {
     win = env.win;
     doc = win.document;
     StubElementClass = doc.registerElement('amp-stub2', {
-      prototype: createAmpElementProto(win, 'amp-stub2', ElementStub),
+      prototype: createAmpElementProtoForTesting(win, 'amp-stub2', ElementStub),
     });
     env.ampdoc.declareExtension_('amp-stub2');
     element = new StubElementClass();
@@ -1597,8 +1462,10 @@ describes.realWin('CustomElement Loading Indicator', {amp: true}, env => {
     doc = win.document;
     clock = lolex.install(win);
     ElementClass = doc.registerElement('amp-test-loader', {
-      prototype: createAmpElementProto(win, 'amp-test-loader', TestElement),
+      prototype: createAmpElementProtoForTesting(
+          win, 'amp-test-loader', TestElement),
     });
+    win.ampExtendedElements['amp-test-loader'] = TestElement;
     LOADING_ELEMENTS_['amp-test-loader'.toUpperCase()] = true;
     resources = Services.resourcesForDoc(doc);
     resources.isBuildOn_ = true;
@@ -1931,7 +1798,8 @@ describes.realWin('CustomElement Overflow Element', {amp: true}, env => {
     win = env.win;
     doc = win.document;
     ElementClass = doc.registerElement('amp-test-overflow', {
-      prototype: createAmpElementProto(win, 'amp-test-overflow', TestElement),
+      prototype: createAmpElementProtoForTesting(
+          win, 'amp-test-overflow', TestElement),
     });
     resources = Services.resourcesForDoc(doc);
     resourcesMock = sandbox.mock(resources);
@@ -2015,139 +1883,5 @@ describes.realWin('CustomElement Overflow Element', {amp: true}, env => {
 
     expect(overflowElement.onclick).to.not.exist;
     expect(overflowElement).to.not.have.class('amp-visible');
-  });
-
-  describe('no body', () => {
-
-    let elements;
-    let doc;
-    let win;
-    let elem1;
-    let ampdoc;
-
-    beforeEach(() => {
-      elements = [];
-
-      doc = {
-        registerElement: sandbox.spy(),
-        documentElement: {
-          ownerDocument: doc,
-        },
-        head: {
-          nodeType: /* ELEMENT */ 1,
-          querySelectorAll: selector => {
-            if (selector == 'script[custom-element]') {
-              return elements;
-            }
-            return [];
-          },
-        },
-        body: {},
-      };
-
-      elem1 = {
-        getAttribute: name => {
-          if (name == 'custom-element') {
-            return 'amp-test1';
-          }
-        },
-        ownerDocument: doc,
-      };
-      elements.push(elem1);
-
-      win = {
-        document: doc,
-        Object: {
-          create: proto => Object.create(proto),
-        },
-        HTMLElement,
-        ampExtendedElements: {},
-      };
-      doc.defaultView = win;
-
-      ampdoc = new AmpDocSingle(win);
-
-      installDocumentStateService(win);
-    });
-
-    afterEach(() => {
-      resetScheduledElementForTesting(win, 'amp-test1');
-      resetScheduledElementForTesting(win, 'amp-test2');
-    });
-
-    it('should be stub elements when body available', () => {
-      stubElementsForDoc(ampdoc);
-
-      expect(win.ampExtendedElements).to.exist;
-      expect(win.ampExtendedElements['amp-test1']).to.equal(ElementStub);
-      expect(win.ampExtendedElements['amp-test2']).to.be.undefined;
-      expect(doc.registerElement).to.be.calledOnce;
-      expect(doc.registerElement.firstCall.args[0]).to.equal('amp-test1');
-    });
-
-    it('should repeat stubbing when body is not available', () => {
-      doc.body = null;  // Body not available
-
-      stubElementsForDoc(ampdoc);
-
-      expect(win.ampExtendedElements).to.exist;
-      expect(win.ampExtendedElements['amp-test1']).to.equal(ElementStub);
-      expect(win.ampExtendedElements['amp-test2']).to.be.undefined;
-      expect(doc.registerElement).to.be.calledOnce;
-      expect(doc.registerElement.firstCall.args[0]).to.equal('amp-test1');
-
-      // Add more elements
-      const elem2 = {
-        getAttribute: name => {
-          if (name == 'custom-element') {
-            return 'amp-test2';
-          }
-        },
-        ownerDocument: doc,
-      };
-      elements.push(elem2);
-
-      // Body available. Stub again.
-      doc.body = {};
-      stubElementsForDoc(ampdoc);
-      expect(win.ampExtendedElements['amp-test1']).to.equal(ElementStub);
-      expect(win.ampExtendedElements['amp-test2']).to.equal(ElementStub);
-      expect(doc.registerElement).to.have.callCount(2);
-      expect(doc.registerElement.getCall(1).args[0]).to.equal('amp-test2');
-    });
-
-    it('should stub element when not stubbed yet', () => {
-      // First stub is allowed.
-      stubElementIfNotKnown(win, 'amp-test1');
-
-      expect(win.ampExtendedElements).to.exist;
-      expect(win.ampExtendedElements['amp-test1']).to.equal(ElementStub);
-      expect(doc.registerElement).to.be.calledOnce;
-      expect(doc.registerElement.firstCall.args[0]).to.equal('amp-test1');
-
-      // Second stub is ignored.
-      stubElementIfNotKnown(win, 'amp-test1');
-      expect(doc.registerElement).to.be.calledOnce;
-    });
-
-    it('should copy or stub element definitions in a child window', () => {
-      stubElementIfNotKnown(win, 'amp-test1');
-
-      const registerElement = sandbox.spy();
-      const childWin = {Object, HTMLElement, document: {registerElement}};
-
-      copyElementToChildWindow(win, childWin, 'amp-test1');
-      expect(childWin.ampExtendedElements['amp-test1']).to.equal(ElementStub);
-      const firstCallCount = registerElement.callCount;
-      expect(firstCallCount).to.equal(1);
-      expect(registerElement.getCall(firstCallCount - 1).args[0])
-          .to.equal('amp-test1');
-
-      copyElementToChildWindow(win, childWin, 'amp-test2');
-      expect(childWin.ampExtendedElements['amp-test1']).to.equal(ElementStub);
-      expect(registerElement.callCount > firstCallCount).to.be.true;
-      expect(registerElement.getCall(registerElement.callCount - 1).args[0])
-          .to.equal('amp-test2');
-    });
   });
 });
