+++ test/functional/test-layout.js
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-import {Layout, assertLength, getLengthNumeral, getLengthUnits, parseLength,
+import {Layout, applyStaticLayout,
+    assertLength, getLengthNumeral, getLengthUnits, parseLength,
     parseLayout, assertLengthOrPercent} from '../../src/layout';
-import {applyLayout_} from '../../src/custom-element';
 
 
 describe('Layout', () => {
@@ -155,7 +155,7 @@ describe('Layout', () => {
 
   it('layout=nodisplay', () => {
     div.setAttribute('layout', 'nodisplay');
-    expect(applyLayout_(div)).to.equal(Layout.NODISPLAY);
+    expect(applyStaticLayout(div)).to.equal(Layout.NODISPLAY);
     expect(div.style.width).to.equal('');
     expect(div.style.height).to.equal('');
     expect(div.style.display).to.equal('none');
@@ -169,7 +169,7 @@ describe('Layout', () => {
     div.setAttribute('layout', 'fixed');
     div.setAttribute('width', 100);
     div.setAttribute('height', 200);
-    expect(applyLayout_(div)).to.equal(Layout.FIXED);
+    expect(applyStaticLayout(div)).to.equal(Layout.FIXED);
     expect(div.style.width).to.equal('100px');
     expect(div.style.height).to.equal('200px');
     expect(div).to.have.class('i-amphtml-layout-fixed');
@@ -180,14 +180,14 @@ describe('Layout', () => {
   it('layout=fixed - default with width/height', () => {
     div.setAttribute('width', 100);
     div.setAttribute('height', 200);
-    expect(applyLayout_(div)).to.equal(Layout.FIXED);
+    expect(applyStaticLayout(div)).to.equal(Layout.FIXED);
     expect(div.style.width).to.equal('100px');
     expect(div.style.height).to.equal('200px');
   });
 
   it('layout=fixed - requires width/height', () => {
     div.setAttribute('layout', 'fixed');
-    expect(() => applyLayout_(div)).to.throw(
+    expect(() => applyStaticLayout(div)).to.throw(
         /Expected height to be available/);
   });
 
@@ -195,7 +195,7 @@ describe('Layout', () => {
   it('layout=fixed-height', () => {
     div.setAttribute('layout', 'fixed-height');
     div.setAttribute('height', 200);
-    expect(applyLayout_(div)).to.equal(Layout.FIXED_HEIGHT);
+    expect(applyStaticLayout(div)).to.equal(Layout.FIXED_HEIGHT);
     expect(div.style.width).to.equal('');
     expect(div.style.height).to.equal('200px');
     expect(div).to.have.class('i-amphtml-layout-fixed-height');
@@ -207,7 +207,7 @@ describe('Layout', () => {
     div.setAttribute('layout', 'fixed-height');
     div.setAttribute('height', 200);
     div.setAttribute('width', 'auto');
-    expect(applyLayout_(div)).to.equal(Layout.FIXED_HEIGHT);
+    expect(applyStaticLayout(div)).to.equal(Layout.FIXED_HEIGHT);
     expect(div.style.width).to.equal('');
     expect(div.style.height).to.equal('200px');
     expect(div).to.have.class('i-amphtml-layout-fixed-height');
@@ -220,13 +220,13 @@ describe('Layout', () => {
     div.setAttribute('height', 200);
     div.setAttribute('width', 300);
     expect(function() {
-      applyLayout_(div);
+      applyStaticLayout(div);
     }).to.throw(/Expected width to be either absent or equal "auto"/);
   });
 
   it('layout=fixed-height - default with height', () => {
     div.setAttribute('height', 200);
-    expect(applyLayout_(div)).to.equal(Layout.FIXED_HEIGHT);
+    expect(applyStaticLayout(div)).to.equal(Layout.FIXED_HEIGHT);
     expect(div.style.height).to.equal('200px');
     expect(div.style.width).to.equal('');
   });
@@ -234,14 +234,14 @@ describe('Layout', () => {
   it('layout=fixed-height - default with height and width=auto', () => {
     div.setAttribute('height', 200);
     div.setAttribute('width', 'auto');
-    expect(applyLayout_(div)).to.equal(Layout.FIXED_HEIGHT);
+    expect(applyStaticLayout(div)).to.equal(Layout.FIXED_HEIGHT);
     expect(div.style.height).to.equal('200px');
     expect(div.style.width).to.equal('');
   });
 
   it('layout=fixed-height - requires height', () => {
     div.setAttribute('layout', 'fixed-height');
-    expect(() => applyLayout_(div)).to.throw(
+    expect(() => applyStaticLayout(div)).to.throw(
         /Expected height to be available/);
   });
 
@@ -250,7 +250,7 @@ describe('Layout', () => {
     div.setAttribute('layout', 'responsive');
     div.setAttribute('width', 100);
     div.setAttribute('height', 200);
-    expect(applyLayout_(div)).to.equal(Layout.RESPONSIVE);
+    expect(applyStaticLayout(div)).to.equal(Layout.RESPONSIVE);
     expect(div.style.width).to.equal('');
     expect(div.style.height).to.equal('');
     expect(div).to.have.class('i-amphtml-layout-responsive');
@@ -264,7 +264,7 @@ describe('Layout', () => {
     div.setAttribute('sizes', '50vw');
     div.setAttribute('width', 100);
     div.setAttribute('height', 200);
-    expect(applyLayout_(div)).to.equal(Layout.RESPONSIVE);
+    expect(applyStaticLayout(div)).to.equal(Layout.RESPONSIVE);
     expect(div.style.width).to.equal('');
     expect(div.style.height).to.equal('');
     expect(div).to.have.class('i-amphtml-layout-responsive');
@@ -276,7 +276,7 @@ describe('Layout', () => {
 
   it('layout=fill', () => {
     div.setAttribute('layout', 'fill');
-    expect(applyLayout_(div)).to.equal(Layout.FILL);
+    expect(applyStaticLayout(div)).to.equal(Layout.FILL);
     expect(div.style.width).to.equal('');
     expect(div.style.height).to.equal('');
     expect(div).to.have.class('i-amphtml-layout-fill');
@@ -286,7 +286,7 @@ describe('Layout', () => {
 
   it('layout=container', () => {
     div.setAttribute('layout', 'container');
-    expect(applyLayout_(div)).to.equal(Layout.CONTAINER);
+    expect(applyStaticLayout(div)).to.equal(Layout.CONTAINER);
     expect(div.style.width).to.equal('');
     expect(div.style.height).to.equal('');
     expect(div).to.have.class('i-amphtml-layout-container');
@@ -298,7 +298,7 @@ describe('Layout', () => {
     div.setAttribute('layout', 'flex-item');
     div.setAttribute('width', 100);
     div.setAttribute('height', 200);
-    expect(applyLayout_(div)).to.equal(Layout.FLEX_ITEM);
+    expect(applyStaticLayout(div)).to.equal(Layout.FLEX_ITEM);
     expect(div.style.width).to.equal('100px');
     expect(div.style.height).to.equal('200px');
     expect(div).to.have.class('i-amphtml-layout-flex-item');
@@ -309,14 +309,14 @@ describe('Layout', () => {
   it('layout=unknown', () => {
     div.setAttribute('layout', 'foo');
     expect(function() {
-      applyLayout_(div);
+      applyStaticLayout(div);
     }).to.throw(/Unknown layout: foo/);
   });
 
 
   it('should configure natural dimensions; default layout', () => {
     const pixel = document.createElement('amp-pixel');
-    expect(applyLayout_(pixel)).to.equal(Layout.FIXED);
+    expect(applyStaticLayout(pixel)).to.equal(Layout.FIXED);
     expect(pixel.style.width).to.equal('0px');
     expect(pixel.style.height).to.equal('0px');
   });
@@ -324,7 +324,7 @@ describe('Layout', () => {
   it('should configure natural dimensions; default layout; with width', () => {
     const pixel = document.createElement('amp-pixel');
     pixel.setAttribute('width', '11');
-    expect(applyLayout_(pixel)).to.equal(Layout.FIXED);
+    expect(applyStaticLayout(pixel)).to.equal(Layout.FIXED);
     expect(pixel.style.width).to.equal('11px');
     expect(pixel.style.height).to.equal('0px');
   });
@@ -332,7 +332,7 @@ describe('Layout', () => {
   it('should configure natural dimensions; default layout; with height', () => {
     const pixel = document.createElement('amp-pixel');
     pixel.setAttribute('height', '11');
-    expect(applyLayout_(pixel)).to.equal(Layout.FIXED);
+    expect(applyStaticLayout(pixel)).to.equal(Layout.FIXED);
     expect(pixel.style.width).to.equal('0px');
     expect(pixel.style.height).to.equal('11px');
   });
@@ -340,7 +340,7 @@ describe('Layout', () => {
   it('should configure natural dimensions; layout=fixed', () => {
     const pixel = document.createElement('amp-pixel');
     pixel.setAttribute('layout', 'fixed');
-    expect(applyLayout_(pixel)).to.equal(Layout.FIXED);
+    expect(applyStaticLayout(pixel)).to.equal(Layout.FIXED);
     expect(pixel.style.width).to.equal('0px');
     expect(pixel.style.height).to.equal('0px');
   });
@@ -348,7 +348,7 @@ describe('Layout', () => {
   it('should configure natural dimensions; layout=fixed-height', () => {
     const pixel = document.createElement('amp-pixel');
     pixel.setAttribute('layout', 'fixed-height');
-    expect(applyLayout_(pixel)).to.equal(Layout.FIXED_HEIGHT);
+    expect(applyStaticLayout(pixel)).to.equal(Layout.FIXED_HEIGHT);
     expect(pixel.style.height).to.equal('0px');
     expect(pixel.style.width).to.equal('');
   });
@@ -360,26 +360,26 @@ describe('Layout', () => {
     pixel.setAttribute('width', '1px');
     pixel.setAttribute('height', '1px');
     expect(() => {
-      applyLayout_(pixel);
+      applyStaticLayout(pixel);
     }).to.not.throw();
 
     // Width=auto is also correct.
     pixel.setAttribute('width', 'auto');
     expect(() => {
-      applyLayout_(pixel);
+      applyStaticLayout(pixel);
     }).to.not.throw();
 
     // Width=X is invalid.
     pixel.setAttribute('width', 'X');
     expect(() => {
-      applyLayout_(pixel);
+      applyStaticLayout(pixel);
     }).to.throw(/Invalid width value/);
 
     // Height=X is invalid.
     pixel.setAttribute('height', 'X');
     pixel.setAttribute('width', '1px');
     expect(() => {
-      applyLayout_(pixel);
+      applyStaticLayout(pixel);
     }).to.throw(/Invalid height value/);
   });
 
@@ -391,7 +391,7 @@ describe('Layout', () => {
     div.className = 'other';
     div.style.width = '111px';
     div.style.height = '112px';
-    expect(applyLayout_(div)).to.equal(Layout.FLEX_ITEM);
+    expect(applyStaticLayout(div)).to.equal(Layout.FLEX_ITEM);
     // No other attributes are read or changed.
     expect(div.style.width).to.equal('111px');
     expect(div.style.height).to.equal('112px');
@@ -404,26 +404,26 @@ describe('Layout', () => {
     div.setAttribute('i-amphtml-layout', 'responsive');
     const sizer = document.createElement('i-amphtml-sizer');
     div.appendChild(sizer);
-    expect(applyLayout_(div)).to.equal(Layout.RESPONSIVE);
-    expect(div.sizerElement_).to.equal(sizer);
+    expect(applyStaticLayout(div)).to.equal(Layout.RESPONSIVE);
+    expect(div.sizerElement).to.equal(sizer);
   });
 
   it('should allow sizer to be missing', () => {
     div.setAttribute('i-amphtml-layout', 'responsive');
-    expect(applyLayout_(div)).to.equal(Layout.RESPONSIVE);
-    expect(div.sizerElement_).to.be.undefined;
+    expect(applyStaticLayout(div)).to.equal(Layout.RESPONSIVE);
+    expect(div.sizerElement).to.be.undefined;
   });
 
   it('should allow sizer to be missing even if other children there', () => {
     div.setAttribute('i-amphtml-layout', 'responsive');
     const other = document.createElement('div');
     div.appendChild(other);
-    expect(applyLayout_(div)).to.equal(Layout.RESPONSIVE);
-    expect(div.sizerElement_).to.be.undefined;
+    expect(applyStaticLayout(div)).to.equal(Layout.RESPONSIVE);
+    expect(div.sizerElement).to.be.undefined;
   });
 
   it('should fail when server generates invalid layout', () => {
     div.setAttribute('i-amphtml-layout', 'invalid');
-    expect(() => applyLayout_(div)).to.throw(/failed/);
+    expect(() => applyStaticLayout(div)).to.throw(/failed/);
   });
 });
